GreenFunc

Documentation for GreenFunc.

GreenFunc.Green2DLR

GreenFunc.toTau

GreenFunc.Green2DLR — Type

Green's function with two external legs that has in-built Discrete Lehmann Representation.

#Members:

'name': Name of green's function

'isFermi': Particle is fermi or boson

'β': Inverse temperature

'timeType': Whether the Green's function is in time/frequency/dlr space

'timeSymmetry': Whether the Green's function has particle-hole symmetry, anti-particle-hole symmetry or none of them

'timeGrid': Time or Frequency grid

'spaceType': Whether the Green's function is in coordinate space/momentum space

'spaceSymmetry': Symmetry of lattice

'spaceGrid': Coordinate or momentum grid

'color': Indices of species of Green's function (such as different spin values)

'dlrGrid': In-built Discrete Lehmann Representation

'instant': Instantaneous part of Green's function that is proportional to δ(τ) in τ space.

'dynamic': Dynamic part of Green's function

'instantError': Error of instantaneous part

'dynamicError': Error of dynamic part

GreenFunc.toTau — Function

function toTau(green::Green2DLR, targetGrid =  green.dlrGrid.τ)
Convert Green's function to τ space by Fourier transform.
If green is already in τ space then it will be interpolated to the new grid.
#Arguements
- 'green': Original Green's function
- 'targetGrid': Grid of outcome Green's function. Default: DLR τ grid