Green's Functions

Base.:<< — Method

Base.:<<(objL::MeshArray, objR::MeshArray)

DLR Fourier transform of functions on the first temporal grid (ImTime, ImFreq or DLRFreq).

If objL and objR have identical temporal grid, objL<<objR assign objR to objL.

If objL and objR have different temporal grid, one of them has to be in DLR space.

If objL is in DLR space, objL<<objR calculates the DLR spectral density of data in objR

if objR is in DLR space, objL<<objR calculates the Green's function from the DLR spectral density in objR.

GreenFunc.dlr_to_imtime — Method

function dlr_to_imfreq(mesharray[, tgrid; dim])

Transform a Green's function in DLR to the imaginary-time domain. #Arguements

'mesharray': MeshArray in DLR space

tgrid: The imaginary-time grid which the function transforms into. Default value is the imaginary-time grid from the DLRGrid from mesharray.mesh[dim].

dim: The dimension of the temporal mesh. Default value is the first ImTime mesh.

GreenFunc.imfreq_to_dlr — Method

function imfreq_to_dlr(mesharray[; dim])

Calculate the DLR sepctral density of a Matsubara-frequency Green's function. #Arguements

'mesharray': MeshArray in the Matsubara-frequency domain.

dim: The dimension of the mesh to be transformed. Default value is the first dimension with mesh type ImFreq.

GreenFunc.imtime_to_dlr — Method

function imtime_to_dlr(mesharray[; dim])

Calculate the DLR sepctral density of an imaginary-time Green's function.

#Arguements

'mesharray': MeshArray in the imaginary-time domain.

dim: The dimension of the mesh to be transformed. Default value is the first dimension with mesh type ImTime.

GreenFunc.to_dlr — Method

function to_dlr(mesharray[; dim])

Calculate the DLR sepctral density of an imaginary-time or Matsubara-frequency Green's function.

#Arguements

'mesharray': MeshArray in the imaginary-time or the Matsubara-frequency domain.

dim: The dimension of the mesh to be transformed. Default value is the first dimension with mesh type ImTime or ImFreq.

GreenFunc.to_imfreq — Method

function to_imfreq(mesharray[; dim])

Transform a Green's function to the Matsubara-frequency domain.

#Arguements

'mesharray': MeshArray in the imaginary-time, the Matsubara-frequency or the DLR frequency domain.

dim: The dimension of the mesh to be transformed. Default value is the first dimension with mesh type DLRFreq, ImTime or ImFreq.

GreenFunc.to_imtime — Method

function to_imtime(mesharray[; dim])

Transform a Green's function to the imaginary-time domain.

#Arguements

'mesharray': MeshArray in the imaginary-time, the Matsubara-frequency or the DLR frequency domain.

dim: The dimension of the mesh to be transformed. Default value is the first dimension with mesh type DLRFreq, ImTime or ImFreq.