Green's Functions

GreenFunc.dlr_to_imfreq — Method

function dlr_to_imfreq(obj::MeshArray, ngrid=nothing; dim=nothing)

Transform a Green's function in DLR space to Matsubara frequency space. #Arguements

'obj': Function in DLR space

ngrid: The Matsubara-frequency grid which the function transforms into. Default value is the Matsubara-frequency grid from the ImFreq constructor.

dim: The dimension of the temporal mesh. Default value is the first ImFreq mesh.

GreenFunc.dlr_to_imtime — Method

function dlr_to_imtime(obj::MeshArray, tgrid=nothing; dim=nothing)

Transform a Green's function in DLR space to the imaginary-time space. #Arguements

'obj': Function in DLR space

tgrid: The imaginary-time grid which the function transforms into. Default value is the imaginary-time grid from the ImFreq constructor.

dim: The dimension of the temporal mesh. Default value is the first ImTime mesh.

GreenFunc.imfreq_to_dlr — Method

function imfreq_to_dlr(obj::MeshArray; dim=nothing, rtol=1e-12, sym=:none)

Calculate the DLR sepctral density of a Matsubara-frequency Green's function. #Arguements

'obj': Function in the Matsubara-frequency space

dim: The dimension of the mesh to be transformed. Default value is the first dimension with mesh type ImFreq.

rtol: The relative tolerance of the DLR transform. Default value is 1e-12.

sym: The symmetry of the Green's function, :none, :ph or :pha. Default value is :none.

GreenFunc.imtime_to_dlr — Method

function imtime_to_dlr(obj::MeshArray; dim=nothing, rtol=1e-12, sym=:none)

Calculate the DLR sepctral density of an imaginary-time Green's function. #Arguements

'obj': Function in the imaginary-time space

dim: The dimension of the mesh to be transformed. Default value is the first dimension with mesh type ImTime.

rtol: The relative tolerance of the DLR transform. Default value is 1e-12.

sym: The symmetry of the Green's function, :none, :ph or :pha. Default value is :none.

GreenFunc.to_dlr — Method

function to_dlr(obj::MeshArray; dim=nothing, rtol=1e-12, sym=:none)

Calculate the DLR sepctral density of an imaginary-time or Matsubara-frequency Green's function. #Arguements

'obj': Function in the imaginary-time space or in the Matsubara-frequency

dim: The dimension of the mesh to be transformed. Default value is the first dimension with mesh type ImTime or ImFreq.

rtol: The relative tolerance of the DLR transform. Default value is 1e-12.

sym: The symmetry of the Green's function, :none, :ph or :pha. Default value is :none.