MeshGrids

GreenFunc.MeshGrids.DLRFreq — Type

function DLRFreq(β, isFermi::Bool=false;
    dtype=Float64,
    rtol=1e-12,
    Euv=1000 / β,
    sym=:none,
    rebuild=false,
    dlr::Union{DLRGrid,Nothing}=nothing
)

Create a DLRFreq struct.

Arguments

β: inverse temperature.

isFermi: the statistics for particles is fermionic or not. False by default.

dtype: type of β and Euv.

rtol: tolerance absolute error. By default, rtol = 1e-12.

Euv: the UV energy scale of the spectral density. By default, Euv = 1000 / β.

sym: the symmetry of dlr. By default, sym = :none.

rebuild: if no dlr is input, set false to load DLRGrid from the file; set true to recalculate the DLRGrid on the fly. By default, rebuild = false.

dlr: 1D DLR grid. By default, a DLR grid with input arguments is used.

GreenFunc.MeshGrids.DLRFreq — Type

struct DLRFreq{T<:Real} <: TemporalGrid{Int}

Discrete-Lehmann-representation grid for Green's functions.

Parameters

T: type of the grid point, β and Euv.

Members

dlr: built-in DLR grid.

grid: 1D grid of time axis, with locate, volume, and AbstractArray interface implemented. It should be grid of Int for ImFreq, and DLRGrid for DLRFreq.

β: inverse temperature.

Euv: the UV energy scale of the spectral density.

rtol: tolerance absolute error.

sym: the symmetry of dlr.

isFermi: the statistics for particles.

GreenFunc.MeshGrids.ImFreq — Type

function ImFreq(β, isFermi::Bool=false;
    dtype=Float64,
    Euv=1000 / β,
    grid::Union{AbstractGrid,AbstractVector,Nothing}=nothing
)

Create a ImFreq struct.

Arguments

β: inverse temperature.

isFermi: the statistics for particles is fermionic or not. False by default.

dtype: type of β and Euv. By default, dtype = Float64.

Euv: the UV energy scale of the spectral density. By default, Euv = 1000 / β.

grid: 1D Matsubara-frequency integer-valued grid as a AbstractVector or CompositeGrids.AbstractGrid. By default, a optimized grid built in DLR is used.

GreenFunc.MeshGrids.ImFreq — Type

struct ImFreq{T<:Real, Grid} <: TemporalGrid{Int}

Imaginary-frequency grid for Green's functions.

Parameters

T: type of β and Euv.

Grid: type of 1D grid with Int as the grid point type.

Members

grid: 1D grid of time axis, with locate, volume, and AbstractArray interface implemented. It should be grid of Int for ImFreq, and DLRGrid for DLRFreq.

β: inverse temperature.

Euv: the UV energy scale of the spectral density.

isFermi: the statistics for particles is fermionic or not.

GreenFunc.MeshGrids.ImTime — Type

function ImTime(β, isFermi::Bool=false;
    dtype=Float64,
    rtol=1e-12,
    Euv=1000 / β,
    grid::Union{AbstractGrid,AbstractVector,Nothing}=nothing
)

Create a ImTime struct.

Arguments

β: inverse temperature.

isFermi: the statistics for particles is fermionic or not. False by default.

dtype: type of the grid point. By default, dtype = Float64.

Euv: the UV energy scale of the spectral density. By default, Euv = 1000 / β.

grid: 1D time grid as a AbstractVector or CompositeGrids.AbstractGrid. By default, a optimized grid built in DLR is used.

GreenFunc.MeshGrids.ImTime — Type

struct ImTime{T, Grid} <: TemporalGrid{T}

Time grid for Green's functions.

Parameters

T: type of the grid point, β and Euv.

Grid: type of 1D grid with T as the grid point type.

Members

grid: 1D grid of time axis, with locate, volume, and AbstractArray interface implemented. It should be grid of Int for ImFreq, and DLRGrid for DLRFreq.

β: inverse temperature.

Euv: the UV energy scale of the spectral density.

isFermi: the statistics for particles is fermionic or not.

GreenFunc.MeshGrids.MeshProduct — Type

The cartesian Mesh product:

#Parameters:

'MT': Type of meshes

'N' : Number of meshes

#Members:

'meshes' : The list of Meshes in the MeshProduct

'dims' : A tuple of the length of the mesh factors

Base.getindex — Method

getindex(g::ImFreq, I::Int)

Equivalent to g[I], get the real-valued Matsubara frequency of the Ith point in the grid. For fermion, return (2g[I]+1)π/β, for boson, return 2g[I]*π/β.

If you need the integer-valued frequency, use g.grid[I] instead.

Base.length — Method

function Base.length(obj::MeshProduct)

Return the number of grids of the MeshProduct.

Base.show — Method

show(io::IO, tg::DLRFreq)

Write a text representation of the DLR grid tg to the output stream io.

Base.show — Method

show(io::IO, tg::ImFreq)

Write a text representation of the Imaginary-frequency grid tg to the output stream io.

Base.show — Method

show(io::IO, tg::ImTime)

Write a text representation of the Imaginary-time grid tg to the output stream io.

Base.show — Method

function Base.show(io::IO, obj::MeshProduct)

Print the MeshProduct.

Base.size — Method

function Base.size(obj::MeshProduct, I::Int)

Return the length of the specifict Ith mesh factor of the MeshProduct.

Base.size — Method

function Base.size(obj::MeshProduct, I::Int)

Return the length of the specifict Ith mesh factor of the MeshProduct.

GreenFunc.MeshGrids.index_to_linear — Method

function index_to_linear(obj::MeshProduct, index...)

Convert a tuple of the indexes of each mesh to a single linear index of the MeshProduct.

Argument:

'obj': The MeshProduct object

'index...': N indexes of the mesh factor, where N is the number of mesh factor

GreenFunc.MeshGrids.linear_to_index — Method

function linear_to_index(obj::MeshProduct, I::Int)

Convert the single linear index of the MeshProduct to a tuple of indexes of each mesh.

Argument:

'obj': The MeshProduct object

'I': The linear index of the MeshProduct

GreenFunc.MeshGrids.locate — Method

locate(tg::ImFreq, n::Int)
locate(tg::ImFreq, ωn)

Find the location in tg.grid for the Matsubara frequency ωn or the integer n.